requirements.txt
accelerate==0.29.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pymongo==4.6.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
urllib3==2.2.1
